renaming todo files to backlogItem
