+++ backend/src/router/backlogItem.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import {createStory, editStory} from '../controllers/todo';
+import {createStory, editStory} from '../controllers/backlogItem';
 
 export default(router: express.Router) => {
   router.post('sprint/:sprintId/create', createStory);
